pyproject.toml
[project]
name = "rl-core"
version = "0.1.0"
description = "Shared infrastructure for reinforcement learning research"
authors = [
    {name = "eoinmcallister",email = "eoinmc.james@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0",
    "numpy>=1.26",
    "pyyaml>=6.0",
    "wandb>=0.15",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
ruff = "^0.15.8"
ty = "^0.0.26"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = []

[tool.ruff.format]
quote-style = "double"

